% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suvModel_functions.R
\name{build_pairwise_logrank}
\alias{build_pairwise_logrank}
\title{function to compare logrank pairwise}
\usage{
build_pairwise_logrank(
  x,
  variables = x[["descriptive_columns"]],
  strata_variables = NULL,
  group_by = NULL,
  pivot_wider = FALSE,
  ...
)
}
\description{
function to compare logrank pairwise
}
