% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suvModel_functions.R
\name{build_surv_model_single_}
\alias{build_surv_model_single_}
\title{function to build model}
\usage{
build_surv_model_single_(
  x,
  variables = x[["descriptive_columns"]],
  strata_variables = NULL,
  model_type = c("survfit", "logrank", "coxph"),
  ...
)
}
\description{
function to build model
}
