% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suvModel_functions.R
\name{compare_levels}
\alias{compare_levels}
\title{function to build model}
\usage{
compare_levels(
  x,
  variables = x[["descriptive_columns"]],
  strata_variables = NULL,
  p.adjust.method = "BH"
)
}
\description{
function to build model
}
