% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_functions.R
\name{create_surv_formula}
\alias{create_surv_formula}
\title{function to create surv formula}
\usage{
create_surv_formula(x, variables = NULL, strata_variables = NULL)
}
\description{
function to create surv formula
}
