% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1_functions.R
\name{create_table1}
\alias{create_table1}
\title{function to create surv formula}
\usage{
create_table1(
  x,
  variables = x[["descriptive_columns"]],
  strata_variables = NULL,
  add_pvalue = TRUE,
  overall = "Overall",
  ...
)
}
\description{
function to create surv formula
}
