% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_functions.R
\name{create_table_formula}
\alias{create_table_formula}
\title{function to create table1 formula}
\usage{
create_table_formula(
  x,
  variables = x[["descriptive_columns"]],
  strata_variables = NULL
)
}
\description{
function to create table1 formula
}
