% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suvModel_functions.R
\name{extract_signif_vars_cox}
\alias{extract_signif_vars_cox}
\title{function to build model}
\usage{
extract_signif_vars_cox(
  x,
  variables = x[["descriptive_columns"]],
  strata_variables = NULL,
  group_by = NULL,
  thershold = 0.05,
  ...
)
}
\description{
function to build model
}
