% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_functions.R
\name{get_variables_surv_formula}
\alias{get_variables_surv_formula}
\title{function to create table1 formula}
\usage{
get_variables_surv_formula(surv_formula)
}
\description{
function to create table1 formula
}
