% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_functions.R
\name{get_variables_surv_model}
\alias{get_variables_surv_model}
\title{function to create table1 formula}
\usage{
get_variables_surv_model(surv_model)
}
\description{
function to create table1 formula
}
