% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{makeSurvReport}
\alias{makeSurvReport}
\title{generate report rmd}
\usage{
makeSurvReport(
  x,
  output_file,
  single_variables,
  main_variable = NULL,
  group_variables = NULL,
  strata_variables = NULL,
  OS_rate_timepoint = 5 * 365.25,
  path_to_script = NULL,
  ignored_variables = NULL,
  xlsx_path = NULL,
  create_model = TRUE
)
}
\description{
generate report rmd
}
