% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survData.R
\name{new_survData}
\alias{new_survData}
\title{new class}
\usage{
new_survData(
  data,
  time_column,
  event_column,
  descriptive_columns,
  factor_columns,
  numeric_columns,
  weights
)
}
\arguments{
\item{data}{A number.}

\item{time_column}{A number.}

\item{event_column}{event}
}
\description{
new class
}
