% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity_score.R
\name{propensity_matching_single_}
\alias{propensity_matching_single_}
\title{function to propensity score matching}
\usage{
propensity_matching_single_(
  data,
  vars_to_matching,
  var_treat,
  control_indicator = 0,
  ps_method = "matching",
  ...
)
}
\description{
function to propensity score matching
}
