% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survData.R
\name{survData}
\alias{survData}
\title{new class}
\usage{
survData(
  data,
  time_column,
  event_column,
  descriptive_columns = NULL,
  weights = NULL
)
}
\arguments{
\item{data}{A number.}

\item{time_column}{A number.}

\item{event_column}{event}

\item{descriptive_columns}{variables to describe}
}
\description{
new class
}
